Refactor WebSockets in Redux
Move ws middleware and connection to `ws` subdir

Rename `wsActions.sendWSGet` to `wsActions.sendGet`

Move `wsActions` to the separate file

Rename chat file to import `@/store/chats` instead of `@/store/chats/chats`
