+++ frontend/src/app/chat/[id]/chat-page.tsx
@@ -11,7 +11,7 @@ import Markdown from "@/components/markdown/markdown"
 import { useChatMessages, useMessage, useSendMessage } from "@/api/chats/messages"
 import { useIsUserOnline, useUser } from "@/api/users/hooks"
 import { useChat } from "@/api/chats/chats"
-import { chatActions } from "@/store/chats/chats"
+import { chatActions } from "@/store/chats"
 import { useMe } from "@/api/auth"
 
 const ChatPage = () => {
