+++ frontend/src/store/ws/actions.ts
@@ -1,0 +1,35 @@
+import { createAction } from "@reduxjs/toolkit"
+
+export const sendGet = createAction("ws/send", (url: string) => {
+  return {
+    payload: {
+      type: "get",
+      item: {
+        url,
+      },
+    },
+  }
+})
+
+// TODO: improve types, maybe even infer data type by url
+export const sendPost = createAction("ws/send", <T>(url: string, data: T) => {
+  return {
+    payload: {
+      type: "post",
+      item: {
+        url,
+        data,
+      },
+    },
+  }
+})
+
+const close = createAction("ws/close")
+
+const wsActions = {
+  sendGet,
+  sendPost,
+  close,
+}
+
+export default wsActions
