+++ frontend/src/store/ws/connection.ts
